+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # changelog
 
 * add `DelRuneInStringWith` and `FilterRuneInStringWith` functions in `basicalter` package (remove specific rune(s) on the value of a string pointer with functions)
+* avoid panic when `find` is nil with function `OneInSliceWith` in `basiccheck` package
 
 ## v0.8.0